+++ src/remote/valkey-glide/index.ts
@@ -54,11 +54,15 @@ export class ValkeyGlideCache extends BaseCache {
   }
 
   override async getMany<T>(keys: string[]): Promise<Record<string, T>> {
+    if (keys.length === 0) {
+      return {};
+    }
+
     this.logger?.debug(this.name, '[getMany] Running "MGET" command...', 'keys =', keys);
 
     const values = await this.client.mget(keys);
 
-    const data: Record<string, T> = {};
+    const data: Record<string, T> = Object.create(null);
 
     for (let i = 0; i < keys.length; i++) {
       const value = values[i];
@@ -72,6 +76,10 @@ export class ValkeyGlideCache extends BaseCache {
   }
 
   override async deleteMany(keys: string[]): Promise<void> {
+    if (keys.length === 0) {
+      return;
+    }
+
     this.logger?.debug(this.name, '[deleteMany] Running "UNLINK" command...', 'keys =', keys);
 
     await this.client.unlink(keys);
